requirements.txt
pyspark
tweepy
python-dotenv